Move Alert/CipherSuite stuff into separate modules, add comment headers
