+++ Sming/Components/ssl/include/Network/Ssl/Certificate.h
@@ -9,6 +9,7 @@
  * @author: 2019 - Slavey Karadzhov <slav@attachix.com>
  *
  ****/
+
 #pragma once
 
 #include <WString.h>
