+++ Sming/Components/ssl/include/Network/Ssl/InputBuffer.h
@@ -1,3 +1,13 @@
+/****
+ * Sming Framework Project - Open Source framework for high efficiency native ESP8266 development.
+ * Created 2015 by Skurydin Alexey
+ * http://github.com/SmingHub/Sming
+ * All files of the Sming Core are provided under the LGPL v3 license.
+ *
+ * InputBuffer.h
+ *
+ ****/
+
 #pragma once
 
 #include <user_config.h>
