+++ Sming/Components/ssl/src/Session.cpp
@@ -1,3 +1,12 @@
+/****
+ * Sming Framework Project - Open Source framework for high efficiency native ESP8266 development.
+ * Created 2015 by Skurydin Alexey
+ * http://github.com/SmingHub/Sming
+ * All files of the Sming Core are provided under the LGPL v3 license.
+ *
+ * Session.cpp
+ *
+ ****/
 
 #include <Network/Ssl/Session.h>
 #include <Network/Ssl/Factory.h>
